refactor: move all of lang into shared #159
