+++ frontend/src/components/AppBar.vue
@@ -15,7 +15,7 @@ You should have received a copy of the GNU Affero General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 -->
 <script setup lang="ts">
-import { language } from "../lang/lang.ts";
+import { language } from "../../../shared/src/lang/lang.ts";
 import { computed, onMounted, ref } from "vue";
 import { useUserStore } from "../store/userStore.ts";
 import { storeToRefs } from "pinia";
