+++ frontend/src/components/AppBarForRegister.vue
@@ -15,7 +15,7 @@ You should have received a copy of the GNU Affero General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 -->
 <script setup lang="ts">
-import { language } from "../lang/lang";
+import { language } from "../../../shared/src/lang/lang";
 </script>
 
 <template>
