+++ frontend/src/components/DepotDialog.vue
@@ -16,7 +16,7 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 -->
 <script setup lang="ts">
 import { computed, inject, Ref, ref } from "vue";
-import { language } from "../lang/lang.ts";
+import { language } from "../../../shared/src/lang/lang.ts";
 import { saveDepot } from "../requests/depot";
 import { Depot, NewDepot, OptionalId } from "../../../shared/src/types.ts";
 const t = language.pages.depots;
